requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
cachetools==5.5.0
celery==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==4.2.18
django-allauth==65.3.0
django-bootstrap-v5==1.0.11
django-celery-beat==2.8.1
django-ckeditor-5==0.2.17
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-defender==0.9.8
django-extensions==4.1
django-filter==24.3
django-jazzmin==2.6.0
django-location-field==2.7.3
django-oauth-toolkit==3.0.1
django-schema-viewer==0.5.3
django-seed==0.3.1
django-simpleui==2025.5.262
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.21.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
exceptiongroup==1.2.2
Faker==33.1.0
filelock==3.16.1
gunicorn==23.0.0
h11==0.16.0
holidays==0.74
httpagentparser==1.9.5
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kombu==5.5.3
msgpack==1.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.90.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg==3.1.18
psycopg-c==3.1.18
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyproject-api==1.8.0
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==1.3.1
rpds-py==0.25.1
service-identity==24.2.0
setuptools==78.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
tomli==2.2.1
toposort==1.10
tox==4.23.2
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
types-PyYAML==6.0.12.20250402
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
upstash-redis==1.4.0
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
